Updated AI to attack the player target
We wanted to finally add the ability to our AI to attack the target when
it locks in. So there were a few changes needed on the c++ side as well
as on the BT side.

- Updated HDCharacter to move StartFire and StopFire to be public
  facing. This is so we could call them from the behaviour tree.
- Added the attacktarget task, that calls StartFire and StopFire.
- Updated behaviour tree, when we're in sequence after TargetActor is
  found. When we're in range of a shot, the AI takes a shot at its
  target. This includes rotating towards the target.

The AI is very accurate. It doesn't look like our recoil is affecting
the AI. So we need to look at some sort of bullet spread next time.