Updates to AI accuracy and spawn
The AI was way too deadly. Not nice. So we added some bullet spread to
the weapon to reduce that deadliness. Of course, now WE are also less
deadly. Also set up the spawner to allow both the bots and Humanoids.

-- Duplicated Rifle to be able to adjust settings differently for our AI
than our player.
-- Changed the game mode to choose, pseudorandomly, if a bot or humanoid
AI is spawned. This is configurable by percent.
-- Added a bulletspread var to the weapon class. We use this to set the
half angle rad of a VRandCone, then use this to set the shotdirection.
This gives us a little random bullet spread.