Added basic teams and friendly fire
Added an int to allow any actor with a health component to have a team.
Then we do an IsFriendly check to ensure we should take damage from the
friendly. Fun enough, if we want to still not allow friendly fire, but
do something like left for dead's "Hey, stop shooting me!" we can do
that from blueprint with the IsFriendly() call.

- Updated health component to keep track of teams through a uint8,
  allowing 256 teams.
- Added IsFriendly() call to health component to check if damaged actor
  and damage causer are on the same team.
- We have to allow self harm, because that's how the bots work
- Updated the weapon class to pass through the owner instead of this as
  the damage causer, otherwise we're passing the weapon which does not
  have health.